Move HeaderTextButton Component to navigation
