+++ src/components/navigation/HeaderTextButton.js
@@ -7,7 +7,7 @@ import { Actions } from 'react-native-router-flux'
 import s from '../../locales/strings.js'
 import { showHelpModal } from '../modals/HelpModal.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
-import { EdgeText } from './EdgeText.js'
+import { EdgeText } from '../themed/EdgeText.js'
 
 type Props = {
   type: 'exit' | 'help'
